% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{add3}
\alias{add3}
\title{Add together two numbers}
\usage{
add3(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A numeric vector.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
```R
add(1,1)
```
```{r}
add(1,10)
```
}
