% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the First Letter of a String}
\usage{
capitalize(x)
}
\arguments{
\item{x}{A character vector of one or more strings.}
}
\value{
A character vector with the first letter of each string capitalized.
}
\description{
Capitalize the First Letter of a String
}
\examples{
capitalize("hello")
capitalize(c("apple", "banana"))
}
