% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassB.R
\name{ClassB}
\alias{ClassB}
\title{R6 Child class of \code{.ClassA}}
\description{
One of the test classes.
}
\section{Super class}{
\code{\link[regexcite:.ClassA]{regexcite::.ClassA}} -> \code{ClassB}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClassB-f1}{\code{ClassB$f1()}}
\item \href{#method-ClassB-f2}{\code{ClassB$f2()}}
\item \href{#method-ClassB-clone}{\code{ClassB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassB-f1"></a>}}
\if{latex}{\out{\hypertarget{method-ClassB-f1}{}}}
\subsection{Method \code{f1()}}{
First test method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassB$f1()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
There is no to return - print some message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassB-f2"></a>}}
\if{latex}{\out{\hypertarget{method-ClassB-f2}{}}}
\subsection{Method \code{f2()}}{
Second test method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassB$f2()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
There is no to return - print some message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClassB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
