% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassC.R
\name{ClassC}
\alias{ClassC}
\title{R6 Child class of \code{.ClassA}}
\description{
One of the test classes.
}
\section{Super class}{
\code{\link[regexcite:.ClassA]{regexcite::.ClassA}} -> \code{ClassC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClassC-f1}{\code{ClassC$f1()}}
\item \href{#method-ClassC-f2}{\code{ClassC$f2()}}
\item \href{#method-ClassC-clone}{\code{ClassC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassC-f1"></a>}}
\if{latex}{\out{\hypertarget{method-ClassC-f1}{}}}
\subsection{Method \code{f1()}}{
First test method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassC$f1()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
There is no to return - print some message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassC-f2"></a>}}
\if{latex}{\out{\hypertarget{method-ClassC-f2}{}}}
\subsection{Method \code{f2()}}{
Second test method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassC$f2()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
There is no to return - print some message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClassC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
