% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMaker.R
\name{ClassMaker}
\alias{ClassMaker}
\title{R6 ClassMaker}
\description{
Creates \code{.ClassB} or \code{.ClassC} with \code{ClassMaker$make()} function.
}
\examples{

## ------------------------------------------------
## Method `ClassMaker$make`
## ------------------------------------------------

class_maker <- ClassMaker$new()
myBClass <- class_maker$make(utils.ClassType$BType)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClassMaker-make}{\code{ClassMaker$make()}}
\item \href{#method-ClassMaker-clone}{\code{ClassMaker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassMaker-make"></a>}}
\if{latex}{\out{\hypertarget{method-ClassMaker-make}{}}}
\subsection{Method \code{make()}}{
Creates \code{.ClassB} or \code{.ClassC}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMaker$make(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{\code{string} class type. It can be "B" or "C"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
New R6Class object if type is valid, otherwise - NULL
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{class_maker <- ClassMaker$new()
myBClass <- class_maker$make(utils.ClassType$BType)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassMaker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClassMaker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassMaker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
