% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassA.R
\name{.ClassA}
\alias{.ClassA}
\title{R6 Abstract class \code{.ClassA}}
\description{
Serves as a base class for classes \code{ClassB} and \code{ClassC}.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-.ClassA-f1}{\code{.ClassA$f1()}}
\item \href{#method-.ClassA-f2}{\code{.ClassA$f2()}}
\item \href{#method-.ClassA-clone}{\code{.ClassA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.ClassA-f1"></a>}}
\if{latex}{\out{\hypertarget{method-.ClassA-f1}{}}}
\subsection{Method \code{f1()}}{
First test method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ClassA$f1()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Error, because this method must be overwritten in child class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.ClassA-f2"></a>}}
\if{latex}{\out{\hypertarget{method-.ClassA-f2}{}}}
\subsection{Method \code{f2()}}{
Second test method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ClassA$f2()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Error, because this method must be overwritten in child class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.ClassA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-.ClassA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ClassA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
