% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_fpds_data.R
\name{get_total_pages}
\alias{get_total_pages}
\title{Get Total Number of Pages}
\usage{
get_total_pages(url)
}
\arguments{
\item{url}{The URL for the page.}
}
\value{
The total number of pages.
}
\description{
This function calculates the total number of pages.
}
