% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{inv_log_quant_transform}
\alias{inv_log_quant_transform}
\title{Inverse Logistic Quantile Transformation}
\usage{
inv_log_quant_transform(value, min, max)
}
\arguments{
\item{max}{}
}
\description{
Inverse Logistic Quantile Transformation
}
