% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{log_quant_transform}
\alias{log_quant_transform}
\title{Logistic Quantile Transformation}
\usage{
log_quant_transform(value, min, max)
}
\arguments{
\item{max}{}
}
\description{
Logistic Quantile Transformation
}
