% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lorem}
\alias{lorem}
\title{lorem example text data}
\format{
\subsection{\code{lorem}}{

A 'lorem ipsum' text of ten joined paragraphs created using 'stringi' package
}
}
\usage{
lorem
}
\description{
A text snippet created using the 'stringi::stri_rand_lipsum()' and a few lines
of string processing
}
\keyword{datasets}
