% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.R
\name{minus}
\alias{minus}
\title{Substract two numeric characters}
\usage{
minus(x, y)
}
\arguments{
\item{x}{Numeric vector of one element}

\item{y}{Numeric vector of one element}
}
\value{
Numeric vector
}
\description{
Substract two numeric characters
}
\examples{
x <- 5
y <- 2
minus(5,2)
}
