% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoralSupport.R
\name{MoralSupport}
\alias{MoralSupport}
\title{MoralSupport}
\usage{
MoralSupport(x)
}
\arguments{
\item{x}{is an integer on a scale from 1 to 10 based on how dumb you're currently feeling}
}
\value{
A heart graph with a message of encouragement
}
\description{
Function produces a message when you input a your current confidence level on a scale from 1 to 10.
}
\examples{
MoralSupport(10)
MoralSupport(3)
MoralSupport(13)


}
