% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A numeric value indicating the first number to be added.}

\item{y}{A numeric value indicating the second number to be added. The default value is 10.}
}
\value{
A numeric value indicating the sum of x and y.
}
\description{
This function takes two parameters, x and y, and returns their sum.
}
\examples{
# Basic usage
my_add(5, 7)

# Using the default value of y
my_add(5)

}
