% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_pkg_data}
\alias{my_pkg_data}
\title{My Package Data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{column1}{Description of column 1}
\item{column2}{Description of column 2}
\item{column3}{Description of column 3}
...
}
}
\usage{
my_pkg_data
}
\description{
A dataset containing the data for my package.
}
\keyword{datasets}
