% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{package-deprecated}
\alias{package-deprecated}
\alias{str_split_one}
\title{Scalar Division}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{x}{The object to be divided}

\item{d}{Scalar to divide \code{x} by.}
}
\value{
The result of the scalar division of \code{x}.
}
\description{
Divide an object by a scalar.
}
\examples{
x <- matrix(1:6, 3L)

}
