% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample  data set}
\format{
\subsection{\code{sample_data}}{

A data.table with 10,056 rows and 6 columns:
\describe{
\item{id}{Unique unit identifer}
\item{site}{Clueter level identifier}
\item{x1, x2, x3}{Predictors of the outcome}
\item{y}{Binary outcome}
}
}
}
\source{
Internally simulated
}
\usage{
sample_data
}
\description{
A simulated data.table
}
\keyword{datasets}
