% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_number_to_list.R
\name{split_number_to_list}
\alias{split_number_to_list}
\title{Split number in characters}
\usage{
split_number_to_list(number)
}
\arguments{
\item{number}{A number of any length}
}
\value{
A character vector.
}
\description{
Split number in characters
}
\examples{
x <- 123
split_number_to_list(x)

y <- 49124
split_number_to_list(y)
}
