% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_contains.R
\name{str_contains}
\alias{str_contains}
\title{Check if a string contains a pattern}
\usage{
str_contains(x, pattern)
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A regular expression pattern to search for.}
}
\value{
A logical vector.
}
\description{
Check if a string contains a pattern
}
\examples{
str_contains("hello world", "world")
str_contains("hello world", "xyz")
}
