% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_count.R
\name{str_count}
\alias{str_count}
\title{Count pattern matches in strings}
\usage{
str_count(string, pattern)
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{A regular expression.}
}
\value{
An integer vector of counts, same length as \code{string}.
}
\description{
Count the number of matches of \code{pattern} within each element of \code{string}.
}
\examples{
str_count("banana", "a")          # 3
str_count(c("abc","baba"), "ba")  # 1, 2
}
