% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_count_matches.R
\name{str_count_matches}
\alias{str_count_matches}
\title{Count the number of pattern matches in a string}
\usage{
str_count_matches(x, pattern)
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A regular expression pattern to count.}
}
\value{
An integer vector.
}
\description{
Count the number of pattern matches in a string
}
\examples{
str_count_matches("banana", "a")
str_count_matches("hello world", "l")
}
