% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_count_words.R
\name{str_count_words}
\alias{str_count_words}
\title{Count the number of words in a string}
\usage{
str_count_words(x, split = " ")
}
\arguments{
\item{x}{A single string}

\item{split}{The pattern to split on, space as default}
}
\value{
An integer giving the number of words
}
\description{
Count the number of words in a string
}
\examples{
str_count_words("hello taiga, you are doing great")
}
