% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect.R
\name{str_detect}
\alias{str_detect}
\title{Detect pattern presence in strings}
\usage{
str_detect(string, pattern)
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{A regular expression (see \link[base:regex]{base::regex}).}
}
\value{
A logical vector, the same length as "string"
}
\description{
A light wrapper around \code{\link[base:grep]{base::grepl()}} that returns a logical vector
indicating whether 'pattern' is found in each element of 'string'
}
\examples{
str_detect("regexcite", "xcite")
str_detect(c("cat", "dog"), "a")

}
