% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_paste.R
\name{str_paste}
\alias{str_paste}
\title{Paste two strings}
\usage{
str_paste(x1, x2 = NULL)
}
\arguments{
\item{x1}{A character vector}

\item{x2}{A character vector to be pasted to the 1st one}
}
\value{
A combined character string of x1 and x2
}
\description{
Paste two strings
}
\examples{
x1 <- "Today is "
x2 <- "a beautiful day!"
str_paste(x1,x2)
}
