% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_reverse.R
\name{str_reverse}
\alias{str_reverse}
\title{Reverse a string}
\usage{
str_reverse(x)
}
\arguments{
\item{x}{A single string.}
}
\value{
A reversed string.
}
\description{
Reverse a string
}
\examples{
str_reverse("desserts")
}
