% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{str_split1}
\alias{str_split1}
\title{Split a string}
\usage{
str_split1(x, split)
}
\arguments{
\item{x}{A character vector with one element}

\item{split}{what to split on}
}
\value{
A character vector
}
\description{
Split a string
}
\examples{
x<-"alpha,beta,charlie,delta"
str_split1(x,",")
}
