% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{str_split_multi}
\alias{str_split_multi}
\title{Split a character vector into a list of character vectors}
\usage{
str_split_multi(string, pattern = ",")
}
\arguments{
\item{string}{Input vector, a character vector, or something coercible to one.}

\item{pattern}{The pattern to look for.}
}
\value{
A list of character vectors
}
\description{
Split a character vector into a list of character vectors
}
\examples{
fruits <- c(
  "apples and oranges and pears and bananas",
   "pineapples and mangos and guavas"
 )
 str_split_multi(fruits, " and ")
}
