% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_subset.R
\name{str_subset}
\alias{str_subset}
\title{Subset string that match a pattern}
\usage{
str_subset(string, pattern)
}
\arguments{
\item{string}{A character vector}

\item{pattern}{A regular expression (see \link[base:regex]{base::regex})}
}
\value{
A character vector containing only elements that match.
}
\description{
A light wrapper around \code{\link[base:grep]{base::grep()}} that returns the matching elements
of "string"
}
\examples{
str_subset(c("cat", "dog", "car"), "^c")
}
