% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_unique}
\alias{str_unique}
\title{Create, modify, and delete columns}
\usage{
str_unique(string, ...)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible
to one. \link{vignette("rd-formatting", package = "roxygen2")}}

\item{...}{Other options used to control matching behavior between duplicate
strings. Passed on to \code{\link[stringi:stri_opts_collator]{stringi::stri_opts_collator()}}. #' Best features of
\code{thisfunction()}:
\itemize{
\item Smells nice
\item Has good vibes
}}
}
\value{
An object of the same type as \code{.data}. The output has the following properties:
\itemize{
\item Rows are a subset of the input, but appear in the same order.
\item Columns are not modified.
\item The number of groups may be reduced (if \code{.preserve} is not \code{TRUE}).
\item Data frame attributes are preserved.
}
}
\description{
\code{mutate()} creates new columns that are functions of existing variables.
It can also modify (if the name is the same as an existing
column) and delete columns (by setting their value to \code{NULL}).
}
\section{Useful mutate functions}{

\itemize{
\item \code{\link{+}}, \code{\link{-}}, \code{\link[=log]{log()}}, etc., for their usual mathematical meanings
}

...
}

\section{Grouped tibbles}{


Because mutating expressions are computed within groups, they may
yield different results on grouped tibbles. This will be the case
as soon as an aggregating, lagging, or ranking function is
involved. Compare this ungrouped mutate:

...
}

\examples{
str_unique(c("a", "b", "c", "b", "a"))
# Use ... to pass additional arguments to stri_unique()
str_unique(c("motley", "mötley", "pinguino", "pingüino"))
str_unique(c("motley", "mötley", "pinguino", "pingüino"), strength = 1)
}
\seealso{
\code{\link[=unique]{unique()}}, \code{\link[stringi:stri_unique]{stringi::stri_unique()}} which this function wraps.
}
