% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_df.R
\name{string_df}
\alias{string_df}
\title{Splits a string and creates an ordered dataframe with the result}
\usage{
string_df(x, split)
}
\arguments{
\item{x}{String to be split.}

\item{split}{Thing to split on.}
}
\value{
A dataframe with two columns: 'order' and 'string'
}
\description{
Splits a string and creates an ordered dataframe with the result
}
\examples{
x <- "alfa,bravo,charlie,delta"
string_df(x, split = ",")
}
