% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Title}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character string to split}

\item{split}{A character pattern to split on.}
}
\value{
A character vector.
}
\description{
Split a string by pattern.
}
\examples{
strsplit1("a,b,c", ",")
}
