% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testData}
\alias{testData}
\title{Text data}
\format{
\subsection{\code{testData}}{

Dataframe with some words
}
}
\source{
\url{https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/cli2.19}
}
\usage{
testData
}
\description{
A template of text for testing the package
}
\keyword{datasets}
