% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{utils.ClassType}
\alias{utils.ClassType}
\title{Class type}
\format{
An object of class \code{list} of length 2.
}
\usage{
utils.ClassType
}
\description{
Its a list with \code{BType} and \code{CType} arguments:
\itemize{
\item \code{BType = "B"}
\item \code{CType = "C"}
}

Used in \link{ClassMaker} class as \code{type} parameter of \code{make} method
}
\concept{utils}
\keyword{datasets}
