% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\docType{package}
\name{yourPackageName-deprecated}
\alias{yourPackageName-deprecated}
\alias{str_split_one}
\title{Deprecated function(s) in the yourPackageName package}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the yourPackageName package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
\code{new_function_name()} \tab now a synonym for \code{\link{str_split_one()}}\cr
}
}

