% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add 2 numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{the first number to be added}

\item{y}{the second number to be added}
}
\value{
the sum of x and y and default y to 10 if not entered
}
\description{
Add 2 numbers
}
\examples{
a <- 1
b <- 2
my_add(a, b)
my_add(a)
}
