% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Addition of 2 inputts}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{first input}

\item{y}{second input}
}
\value{
x + y
}
\description{
Addition of 2 inputts
}
\examples{
x <- 4
y <- 3
my_add(x, y)
}
