% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Sum two values}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{a variable, ideally numeric}

\item{y}{another variable, ideally numeric, with default being 10}
}
\value{
a numeric sum of the two values, if the input values are valid
}
\description{
Sum two values
}
\examples{
my_add(1, 2)
my_add(1)
}
