% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{My add}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{first number}

\item{y}{second number with default value 10}
}
\value{
x + y
}
\description{
My add
}
\examples{

my_add(1, 2)
}
