% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two parameters}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{an integer}

\item{y}{y=10}
}
\value{
integer (x+y)
}
\description{
Add two parameters
}
\examples{
my_add(1, 2)
my_add(1)
}
