% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{the first number to be added}

\item{y}{the second number to be added, default is 10}
}
\value{
NA if one of the input is NA; error message "One of your inputs
contains a string value" if any of the input is not a number
}
\description{
Add two numbers
}
\examples{
my_add(1, 2)
my_add(1)
}
