% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{The first number to add.}

\item{y}{The second number to add (default: 10).}
}
\value{
The sum of x and y.
}
\description{
This function adds two numbers and returns the result.
}
\examples{
my_add(1)    # Returns 11, since y defaults to 10
my_add(1, 2) # Returns 3 , since y is set to 2

}
