% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{My Addition}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{is a number}

\item{y}{is a number, default is 10}
}
\value{
the added value of the inputs x and y
}
\description{
this function adds two parameters, with the default of \code{y} being 10
}
\examples{
my_add(1,2)


}
