% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{The first numeric value to be added}

\item{y}{The Second numeric value to be added.
If it's NA, the default value is 10.}
}
\value{
The sum of x and y.
}
\description{
Add two numbers
}
\examples{
my_add(1)
my_add(1, 2)
my_add(1, NA)
}
