% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A variable.}

\item{y}{A variable with a default at 10.}
}
\value{
The sum of the two variables, NA if either is NA and A warning if One of your inputs contains a string value.
}
\description{
Add two numbers
}
\examples{
my_add(1,11)
my_add(1)
my_add(NA)
my_add(11,NA)
my_add(3,"Hello")
}
