% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{the first number given}

\item{y}{the second number given with default value 10}
}
\value{
the sum of two given numbers
}
\description{
add two numbers
}
\examples{
my_add(1, 2)
}
