% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A number}

\item{y}{A number with default value 10}
}
\value{
A number
}
\description{
Add two numbers
}
\examples{
my_add(1, 2)       # output: 3
my_add(1)          # output: 11
my_add(NA)         # output: NA
my_add(10, NA)     # output: NA
}
