% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add my values}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{First number}

\item{y}{Second numebr with default value 10}
}
\value{
Value added.
}
\description{
Add my values
}
\examples{
my_add(1, 2)
my_add(1)
my_add(NA)
my_add(10, NA)
}
