% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add values}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A number or NA}

\item{y}{A number or NA; default is 10}
}
\value{
sum of x and y
}
\description{
Add values
}
\examples{
my_add(1)
my_add(NA, 1)
my_add(1,3)
}
