% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two number}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A number.}

\item{y}{A number with default value 10.
#' @inheritParams assertthat::is.string}
}
\value{
A number.
}
\description{
Add two number
}
\examples{
x <- 1
y <- 2
my_add(x, y)

}
