% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two parameters}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{An integer}

\item{y}{An integer defaulted at 10}
}
\value{
An integer.
}
\description{
Add two parameters
}
\examples{
x <- 2
y <- 3
my_add(x,y)
}
