% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{An integer}

\item{y}{An optional integer with value 10 if not specified otherwise}
}
\value{
An integer, NA or error message
}
\description{
Add two numbers
}
\examples{
x <- 1
y <- 2
my_add(x,y)
my_add(x)
my_add(99, 103)
}
