% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Sum up two parameters x and y}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{a numeric character}

\item{y}{a numeric character}
}
\value{
A numeric value, sum of values x and y
}
\description{
Sum up two parameters x and y
}
\examples{
my_add(1, 2)
my_add(1)

}
