% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Adds 2 number x and y where y is set to 10 as default}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{Takes input of 1 number.}

\item{y}{Takes input of another number. If none entered it is set to 10.}
}
\value{
A sum for 2 numbers.
}
\description{
Adds 2 number x and y where y is set to 10 as default
}
\examples{
a <- 4
b <- 4
my_add(a, b)
my_add(a)
}
