% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Adds 2 numbers together}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{Integer}

\item{y}{Optional integer. Default value of 10 if no inputs.}
}
\value{
Integer, NA, or Error message
}
\description{
Adds 2 numbers together
}
\examples{
x <- 5
y <- 1
my_add(1, 2)
my_add(x, y)
my_add(x)
}
