% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Title}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A character}

\item{y}{Another character}
}
\value{
A summed character
}
\description{
Title
}
\examples{
x <- 1
y <- 2
my_add(x,y)
}
