% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{The sum of two numerical parameters}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{a numeric parameter or NA}

\item{y}{a numeric parameter or NA; default value is 10}
}
\value{
return NA if at least one parameter is NA, otherwise, return the sum of both parameters
}
\description{
The sum of two numerical parameters
}
\examples{
my_add(1)
my_add(NA, 1)
my_add(1,4)
my_add(NA)

}
