% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Sum of two values}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{One adding number.}

\item{y}{One adding number with default value of 10.}
}
\value{
The output, normally a sum value.
}
\description{
Sum of two values
}
\examples{
x <- 1
y <- 2
my_add(x, y)
}
