% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{package-deprecated}
\alias{package-deprecated}
\alias{str_split_one}
\title{Scalar Division}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with, at most, one element.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described in
\code{vignette("regular-expressions")}. Use \code{\link[stringr:regex]{regex()}} for finer control of the
matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:boundary]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}

\item{n}{Maximum number of pieces to return. Default (Inf) uses all
possible split positions.

For \code{split_split()}, this determines the maximum length of each element
of the output. For \code{str_split_fixed()}, this determines the number of
columns in the output; if an input is too short, the result will be padded
with \code{""}.}

\item{x}{The object to be divided}

\item{d}{Scalar to divide \code{x} by.}
}
\value{
The result of the scalar division of \code{x}.

A character vector.
}
\description{
Divide an object by a scalar.
}
\examples{
x <- matrix(1:6, 3L)

}
