% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexcite-package.R
\docType{package}
\name{regexcite-package}
\alias{regexcite}
\alias{regexcite-package}
\title{regexcite: Make Regular Expressions More Exciting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convenience functions to make some common tasks with string manipulation and regular expressions a bit easier.
}
\author{
\strong{Maintainer}: Luis Barqueira \email{luis.barqueira@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
