% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_comma.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Only splits a string by commas}
\usage{
str_split_comma(string)
}
\arguments{
\item{string}{String to be split}
}
\value{
character vector
}
\description{
Only splits a string by commas
}
\examples{
x <- "a,b,c"
str_split_comma(x)

}
