% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_comma.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a string with commas}
\usage{
str_split_comma(string)
}
\arguments{
\item{string}{A string}
}
\value{
A numeric vector.
}
\description{
Split a string with commas
}
\examples{
str_split_comma("a,b,c")
}
