% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_comma.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a string on commas}
\usage{
str_split_comma(x)
}
\arguments{
\item{x}{A character vector with one element.}
}
\value{
A character vector.
}
\description{
Split a string on commas
}
\examples{
str_split_comma("a,b,c")
}
