% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_comma.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a String by Commas}
\usage{
str_split_comma(string, n = Inf)
}
\arguments{
\item{string}{A character vector with, at most, one element.}

\item{n}{Maximum number of pieces to return. Default (Inf) uses all
possible split positions.

For \code{str_split()}, this determines the maximum length of each element
of the output. For \code{str_split_fixed()}, this determines the number of
columns in the output; if an input is too short, the result will be padded
with \code{""}.}
}
\value{
A list
}
\description{
Split a String by Commas
}
\examples{
str_split_comma("j,a,s")
}
