% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a string on commas}
\usage{
str_split_comma(string)
}
\arguments{
\item{string}{A string to split}
}
\value{
A character vector
}
\description{
Split a string on commas
}
