% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_comma.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a string on commas}
\usage{
str_split_comma(string)
}
\arguments{
\item{string}{A character vector with, at most, one element.}
}
\value{
A character vector.
}
\description{
Splits a single string using commas as the separator.
}
\examples{
str_split_comma("a,b,c")
}
