% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_comma.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a string on commas}
\usage{
str_split_comma(string, n = Inf)
}
\arguments{
\item{string}{A character vector with, at most, one element.}

\item{n}{Maximum number of pieces to return. Default is Inf.}
}
\value{
A character vector.
}
\description{
Split a string on commas
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_comma(x)
str_split_comma(x)

y <- "a,b,c"
str_split_one(y, n = 2)
}
