% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_newfunctions.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Only splits commas}
\usage{
str_split_comma(string)
}
\arguments{
\item{string}{takes a string separated by commas}
}
\value{
character vector
}
\description{
Only splits commas
}
\examples{
str_split_comma("a,b,c")
}
