% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a string with commas}
\usage{
str_split_comma(string, n = Inf)
}
\arguments{
\item{string}{A character vector with, at most, one element.}
}
\value{
A character vector.
}
\description{
Split a string with commas
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_comma(x, pattern = ",")
str_split_comma(x, pattern = ",", n = 2)

y <- "192.168.0.1"
str_split_comma(y)
}
