% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_nums.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Splits string and returns as numeric}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{String to be split}

\item{pattern}{Character for string to be split by}
}
\value{
numeric vector
}
\description{
Splits string and returns as numeric
}
\examples{
x <- "1,2,3"
str_split_comma(x, pattern = ",")

}
