% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_nums.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string and convert to numeric}
\usage{
str_split_nums(x, pattern)
}
\arguments{
\item{x}{A character vector with one element.}

\item{pattern}{What to split on.}
}
\value{
A numeric vector.
}
\description{
Split a string and convert to numeric
}
\examples{
str_split_nums("1,2,3", pattern = ",")
}
