% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_nums.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string and convert to numeric}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{A character vector with one element.}

\item{pattern}{What to split on.}
}
\value{
A numeric vector.
}
\description{
Split a string and convert to numeric
}
