% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string and convert pieces to numbers}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{A string to split}

\item{pattern}{The pattern to split on}
}
\value{
A numeric vector
}
\description{
Split a string and convert pieces to numbers
}
