% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_nums.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string and convert pieces to numbers}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{A character vector with, at most, one element.}

\item{pattern}{Pattern to split on.}
}
\value{
A numeric vector.
}
\description{
Splits a single string using a pattern, then converts the result to numeric values.
}
\examples{
str_split_nums("1,2,3", pattern = ",")
}
