% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_nums.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string and convert to numeric}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{A character vector with, at most, one element.}

\item{pattern}{A string to split on.}
}
\value{
A character vector.
}
\description{
Split a string and convert to numeric
}
\examples{
x <- "1,2,3"
str_split_nums(x, pattern = ",")
}
