% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string and converts it to numeric}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{A character vector with, at most, one element.}

\item{pattern}{A string denoting the separator by which the string parameter
should be split on}
}
\value{
A numeric vector.
}
\description{
Split a string and converts it to numeric
}
\examples{
x <- "1,2,3,4"
str_split_nums(x, pattern = ",")
}
