% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_newfunctions.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Turn number strings to numeric}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{takes a character string of numbers}

\item{pattern}{how the number is seperated}
}
\value{
numeric vector
}
\description{
Turn number strings to numeric
}
\examples{
str_split_nums("1,2,3", ",")
}
