% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string of numbers}
\usage{
str_split_nums(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector of numbers with, at most, one element.}
}
\value{
A numeric vector.
}
\description{
Split a string of numbers
}
\examples{
#' y <- "192.168.0.1"
str_split_nums(y)

x <- "alfa,bravo,charlie,delta"
str_split_nums(x, pattern = ",")
str_split_nums(x, pattern = ",", n = 2)

}
