% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srtsplit1.R
\name{str_split_one}
\alias{str_split_one}
\title{Title}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{hat to split on.}
}
\value{
A character vector.
}
\description{
Title
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
