% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Splitting a string}
\usage{
str_split_one(string, pattern = stringr::fixed("."), n = Inf)
}
\arguments{
\item{string}{Character vector separated by the split character; accept one vector of size 1}

\item{pattern}{Character separator}

\item{n}{Maximum number of pieces to return. Default (Inf) uses all
possible split positions.

For \code{split_split()}, this determines the maximum length of each element
of the output. For \code{str_split_fixed()}, this determines the number of
columns in the output; if an input is too short, the result will be padded
with \code{""}.}
}
\value{
A character vector

character vector
}
\description{
Splitting a string
}
\examples{
x <- c('alpha-beta-gamma-delta')
str_split_one(x, '-')
str_split_one(x, '-', n=2)
}
