% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a String}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{x}{A character vector with one element}

\item{split}{What to spit on}
}
\value{
A character vector
}
\description{
Split a String
}
\examples{
  x <- "alfa,bravo,charlie,delta"
  strsplit1(x, split = ",")

}
