% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{str_split_one: R Regex Function by Jennifer Bryan (Reproduced Here)}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with at most one element, the string to be split}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described in
\code{vignette("regular-expressions")}. Use \code{\link[stringr:regex]{regex()}} for finer control of the
matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:boundary]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}

\item{n}{Maximum number of pieces to return. Default (Inf) uses all
possible split positions.

For \code{split_split()}, this determines the maximum length of each element
of the output. For \code{str_split_fixed()}, this determines the number of
columns in the output; if an input is too short, the result will be padded
with \code{""}.}
}
\value{
the character vector resulting from splitting input 'string' on
delimiter 'pattern' with the returned character vector of maximum length 'n'
}
\description{
str_split_one: R Regex Function by Jennifer Bryan (Reproduced Here)
}
\details{
This function, as well as the full package surrounding it,
are both reproductions from the first chapter of Hadley Wickham and Jennifer
Bryant's R Packages (2e) online book (https://r-pkgs.org/whole-game.html).
The function itself, authored by Jennifer Bryan, is meant to take a
character vector of at most one string 'string' and separate it into parts
delimited by separator 'pattern' with maximum length of the output character
vector defined by parameter 'n'.

The default behavior of the string::str_split function is
to return a list of length 1 containing a character vector of the delimited
elements of the input character vector. regexcite::str_split_one encapsulates
both the call to stringr::str_split() as well as
the common follow-up operation of extracting the desired character vector.

This version of the file has been modified to use stringr instead of base R.
}
\examples{
x <- "Spam,Spam,Spam,Egg,and,Spam"
str_split_one(x, pattern = ",")
str_split_one(x, pattern = ",", n = 2)

y <- "192.168.0.1"
str_split_one(y, pattern = stringr::fixed("."))
}
