% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector}

\item{pattern}{A regex pattern}

\item{n}{An integer}
}
\value{
A character vector
}
\description{
Split a string
}
\examples{
Made by Jason

y <- "192.168.0.1"
str_split_one(y, pattern = stringr::fixed("."))
}
