% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{str_split_one}
\alias{str_split_one}
\title{Split up a string into pieces}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{Input}{vector. Either a character vector, or something coercible to one.}

\item{Pattern}{to look for.}

\item{Maximum}{number of pieces to return. Default (Inf) uses all possible split positions.}
}
\value{
A character vector.
}
\description{
Split up a string into pieces
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_one(string=x, pattern = ",")
}
