% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split string}
\usage{
str_split_one(string, pattern, n = inf)
}
\arguments{
\item{x}{A character vector with element.}

\item{split}{What to split on.}
}
\value{
A character vector one-dimensional.
}
\description{
Split string
}
\examples{
x <- "alpha,bravo,gamma,delta"
strsplit(x, split=",")
}
