% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with element.}

\item{pattern}{What to split on.}

\item{n}{how many elements returned?}
}
\value{
A character vector one-dimensional.
}
\description{
Split string
}
\examples{
x <- "alpha,bravo,gamma,delta"
str_split_one(x, pattern=",")
}
