% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string and return a character vector}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{Input string}

\item{pattern}{What to split by}

\item{n}{Length of each output element}
}
\value{
A character vector
}
\description{
Split a string and return a character vector
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_one(string = x, pattern = ",")
}
