% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with one element.}

\item{pattern}{What to split on}

\item{n}{number of elements}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
# creating a string for testing
x <- "alfa,bravo,charlie,delta"
# applying the function
str_split_one(x, pattern = ",")
}
