% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A single string to be split}

\item{pattern}{regular expresison to split the string}

\item{n}{Maximum number of pieces to return, defaults to Inf (return everything)}
}
\value{
a character vector
}
\description{
Split a string
}
\examples{
str_split_one("ABCD", pattern = "") # c('a','b','c')
}
