% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with one element}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described
in \link[stringi:about_search_regex]{stringi::about_search_regex}. Control options with
\code{\link[stringr:regex]{regex()}}.

For \code{str_replace_all()} this can also be a named vector
(\code{c(pattern1 = replacement1)}), in order to perform multiple replacements
in each element of \code{string}.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.}
}
\value{
A character vector
}
\description{
Split a string
}
\examples{
x <- "alpha,beta,charlie,delta"
str_split_one(x, pattern = ",")
str_split_one(x, pattern = ",", n = 2)
}
