% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string using a pattern, return character vector instead of list}
\usage{
str_split_one(x, split)
}
\arguments{
\item{x}{Character vector with one element}

\item{split}{what to split on}
}
\value{
character vector
}
\description{
Split a string using a pattern, return character vector instead of list
}
\examples{
x <- "alfa,bravo,charlie"
strsplit1(x, split = ",")
}
