% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string into a list of substrings separated by the given separator.}
\usage{
str_split_one(s, sep = ",")
}
\arguments{
\item{s}{A character vector with one element to be split.}

\item{sep}{The separating character to split on. Default: ','}
}
\value{
A character vector of one or more elements
}
\description{
Split a string into a list of substrings separated by the given separator.
}
\examples{
s <- "alfa,bravo,charlie,delta"
str_split_one(s)

s <- "echo:foxtrot:golf:hotel"
str_split_one(s, ':')

s <- "thiswillnotbesplit"
str_split_one(s)
}
