% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{String Split 1}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{Char to be splitted. A char vec of length 1.}

\item{split}{Token to split on. A char vec of length 1.}
}
\value{
A char vec of length 1.
}
\description{
String Split 1
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
