% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{The delimiter to split the string.}
}
\value{
A character vector with the parts of the string.
}
\description{
Split a string
}
\examples{
x <- "a,b,c"
strsplit1(x, split = ",")
}
