% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a long string into a character vector}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{long string}

\item{split}{splitting character}
}
\value{
A character vector
}
\description{
Split a long string into a character vector
}
\examples{
long_string<-("alpha,bravo,charlie,delta")
strsplit1(long_string, ",")
}
