% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Splits string and unlists components}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{String to be split}

\item{split}{character to be split by}
}
\value{
A vector of strings
}
\description{
Splits string and unlists components
}
\examples{
(x <- "alfa,bravo,charlie,delta")
strsplit1(x, split = ",")

}
