% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strisplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{splits the words in a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element}

\item{split}{what to split on}
}
\value{
A character vector
}
\description{
splits the words in a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit(x, split = ","
}
