% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a srtring}
\usage{
strsplit1(x, split)
}
\arguments{
\item{split}{What to split on.}

\item{A}{character vector with one element.}
}
\value{
A character vector.
}
\description{
Split a srtring
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
