% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string once}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector of length 1.}

\item{split}{A character string to split \code{x} on.}
}
\value{
A character vector of the split string.
}
\description{
A helper to split a string into a character vector instead of a list.
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
