% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{strsplit1}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{Character vector separated by the split character}

\item{split}{Split character}
}
\value{
A character vector
}
\description{
strsplit1
}
\examples{
x <- c('eins,zwei,drei')
strsplit1(x, ',')
}
