% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a sting}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element}

\item{split}{What to split on}
}
\value{
A character vector
}
\description{
Split a sting
}
\examples{
x<-"alpha,beta,charlie,delta"
strsplit1(x, split=",")
}
