% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into parts}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A single string to split}

\item{split}{The character to split on}
}
\value{
A character vector of parts
}
\description{
Split a string into parts
}
\examples{
strsplit1("a,b,c", split = ",")
}
