% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into substrings}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector.}

\item{split}{Character string to split \code{x}.}
}
\value{
A character vector of substrings.
}
\description{
Split a string into substrings
}
\examples{
strsplit1("a-b-c", split = "-")
}
