% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{String Manipulation}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{vector}

\item{split}{object to split}
}
\value{
list
}
\description{
String Manipulation
}
\examples{
x <- 'Tall, Short, White, Blue'
strsplit(x, split = ".")
}
