% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Title of the function}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{Character input to split.}

\item{split}{Delimiter to split by.}
}
\value{
A character vector.
}
\description{
Description of what it does.
}
