% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into a vector}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with only one element.}

\item{split}{What to split the string by. Default is a space.}
}
\value{
A character vector containing the split elements.
}
\description{
Split a string into a vector
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
