% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Divide una cadena de texto}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{Un vector de caracter con un elemento.}

\item{split}{Cómo dividir el elemento del vector.}
}
\value{
Un vector de caracter
}
\description{
Divide una cadena de texto
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
