% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A charecter vector with one element.}

\item{split}{A string to split on.}
}
\value{
A character vector
}
\description{
Split a string
}
\examples{
x <- "Oh, say, can, you, see"
strsplit1(x, split = ",")
}
