% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{strsplit1}
\alias{strsplit1}
\title{Title}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{una lista de caracteres}

\item{split}{lo que divide el vector}
}
\value{
char vector
}
\description{
Title
}
\examples{
x <- "alfa,bravo"
strsplit(x,split=",")
}
