% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector of one element}

\item{split}{The separating character used to split x}
}
\value{
The character vector
}
\description{
Split a string
}
\examples{
x <- "cat,dog,hamnster,rabbit"
strsplit1(x, ",")
}
