% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into pieces
simple helper function that splits a single string into its parts, always returns a character vector
it is designed for the common case where the input is "morally" a scalar string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vecto of length one (a single string)}

\item{split}{A character string to use as the delimiter}
}
\value{
A character vector of the split components
}
\description{
Split a string into pieces
simple helper function that splits a single string into its parts, always returns a character vector
it is designed for the common case where the input is "morally" a scalar string
}
\examples{
strsplit1("alfa,bravo,charlie,delta", ",")
}
