% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{string}{A character vector with, at most, one element.}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
strsplit1(x, split = ",")

y <- "192.168.0.1"
strsplit1(y, pattern = stringr::fixed("."))
}
