% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{The character(s) to split on.}
}
\value{
A character Vector.
}
\description{
This function is identical to \code{strsplit()} but returns a \code{vector} rather than a \code{list}.
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
