% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A Character vector}

\item{split}{What to split on}
}
\value{
A Character Vector
}
\description{
Split a string
}
\examples{
x <- "alpha, bravo, charlie, delta"
strsplit1(x, split = ",")
}
