% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{string}{A character vector with, at most, one element.}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_one(x, pattern = ",")
str_split_one(x, pattern = ",", n = 2)

y <- "192.168.0.1"
str_split_one(y, pattern = stringr::fixed("."))
}
