% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a String}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element}

\item{split}{A string to be spitted}
}
\value{
A character vector
}
\description{
Split a String
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
