% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into a character vector}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character string}

\item{split}{A delimiter}
}
\value{
A character vector
}
\description{
This function splits a string using a delimiter and returns a vector.
}
\examples{
strsplit1("dog,cat,bird", ",")
}
