% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split , divide a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{: A character vector}

\item{split}{: split the string}
}
\value{
: A character vector
}
\description{
Split , divide a string
}
\examples{
x <- "alfa, brave, charlie"
strsplit1(x, split = ",")

}
