% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{split a string into a character vector (not a list containing a character vector).}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{A string containing what to split on.}
}
\value{
A character vector.
}
\description{
split a string into a character vector (not a list containing a character vector).
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
