% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into a list of substrings separated by the given separator.}
\usage{
strsplit1(s, sep = ",")
}
\arguments{
\item{s}{A character vector with one element to be split.}

\item{sep}{The separating character to split on. Default: ','}
}
\value{
A character vector of one or more elements
}
\description{
Split a string into a list of substrings separated by the given separator.
}
\examples{
s <- "alfa,bravo,charlie,delta"
strsplit1(s)

s <- "echo:foxtrot:golf:hotel"
strsplit1(s, ':')

s <- "thiswillnotbesplit"
strsplit1(s)
}
