% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into a list of substrings separated by the given separator.}
\usage{
strsplit1(s, sep = ",")
}
\arguments{
\item{s}{The string (i.e. a character vector with one element) to be splited.}

\item{sep}{The separating character to split on. Default: ','}
}
\value{
A character vector
}
\description{
Split a string into a list of substrings separated by the given separator.
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x)

x <- "echo;foxtrot;golf;hotel"
strsplit1(x, ';')
}
